Put logstash indexers on the elasticsearch masters
This requires refactoring the indexer state into a macro so that
different hosts can be specified for the redis and elasticsearch
connections. Which in turn lead to refactoring much of the logstash
state and top file pulling bits into the new state: hpcs. And this
new state seems to be the logical place for the network state to live,
since the state is very hpcloud-specific.

Change-Id: I4ccb4419feebc963c0a1fb7563252cd928b1fb05