That doesn't belong there
Putting an include: statement inside a macro is bad. If the macro
is called multiple times in the same state file, then that entire
state will fail to load into salt and it will be skipped when
compiling highdata for a minion. This means that states included
elsewhere will run, but the state calling the macros will not.

Change-Id: I31cdf05f23b55600269e202a63431b8b95ee2387